# Conflicts:
#	modules/raft/src/main/java/org/apache/ignite/raft/jraft/entity/codec/v1/V1Decoder.java

+++ modules/page-memory/src/main/java/org/apache/ignite/internal/pagememory/util/PartitionlessLinks.java
@@ -74,7 +74,7 @@ public class PartitionlessLinks {
      * @param link Link or page ID.
      * @return Number of bytes written (equal to {@link #PARTITIONLESS_LINK_SIZE_BYTES}).
      */
-    public static long writePartitionless(long addr, long link) {
+    public static int writePartitionless(long addr, long link) {
         putShort(addr, 0, (short) tag(link));
 
         putInt(addr + Short.BYTES, 0, pageIndex(link));
